bpf: Use iptables TPROXY for proxy redirects
Change all Cilium proxies to use transparent redirect (TPROXY). This
allows IP addresses and ports to remain unchanged on redirect4ed
packets and removes the need to use a bpf map to get the original
destination address and port. Thus the proxymap is completely removed,
which helps reduce complexity of the bpf datapath.

Proxies still need to get the source security ID. This is done with a
lookup in the Cilium conntrack bpf map. This only works with the
global conntrack maps, so this commit also deprecates the support for
local CT maps are no longer supported. Much of the code still remains,
but it can not be configured.

Since the bpf socket helper is not yet available, we insert a set of
fixed iptables rules for proxy redirects during Cilium
startup. Proxies themselves can be started later. Due to veth
semantics we still need to modify the packet headers for ingress
proxies, as traffic for those typically arrives via th cilium_host
veth device. We use the DSCP field for this, but only for packets that
are redirected to a host proxy so that the actual endpoints never see
such changes. For egress proxy redirection we use skb mark as egress
traffic always originates from a local Cilium endpoint, and we do not
need to cross a veth pair when redirecting such traffic to a proxy.

Conntrack interferes with TPROXY as conntrack can NAT the traffic due
to 5-tuple collisions while TPROXY may use the non-NATted 5-tuple for
socket matching. This can lead to TCP 3-way handshake never finishing
due to TPROXY directing all packets to the listening socket, instead
of associating non-SYN packets with the new socket. This is resolved
by adding NOTRACK rules for ingress proxy traffic. Egress does not
need this as traffic to egress proxies is always sent from a local
POD, so k8s nodeport NATting does not happen.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>