agent: Wait to regenerate restore endpoints until kvstore is connected
In the process of endpoints being restored from disk, all programs are
regenerated. This bases the programs on the latest snapshots of all local
caches for identities, ipcache, services, and so on. In order to not
destabilize existing endpoint, it must be ensured that local caches have been
synchronized to a valid state. This was done for services as well as identities
if the endpoint was not assigned a fixed or well-known identity. The first
part of the fix is to also require the ipcache to be synchronized to avoid
egress policy drops for endpoints subject to egress policy.

The second part of the fix is to make all endpoints subject to these
synchronization requirements.

Example:
coredns is assigned a well-known identity. When coredns is restored from disk
and regenerated, premature regeneration without the full list of identities
being available will result in the allowed policy map to be stripped of all
global identities. This can result in temporary policy drops.

The risk of avoiding regeneration on restore is that restarting the agent no
longer guarantees to fix an eventual program with the BPF program. It is thus
required for the orchestrator to eventually restart the pod/container on
failure. This will happen in a timely manner with Kubernetes.

Signed-off-by: Thomas Graf <thomas@cilium.io>